requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1

# AI
openai==1.30.1

# Document parsing
pdfplumber==0.11.1
python-docx==1.1.2

# Azure
azure-storage-blob==12.20.0

# Frontend
streamlit==1.35.0

# Utilities
httpx==0.27.0
python-dotenv==1.0.1
aiofiles==23.2.1
